more refactoring and test fixes
