--- lib/validations/url.ts
@@ -1,12 +1,0 @@
-export const validator = (value: string | undefined) => {
-  if (!value) {
-    return true;
-  }
-
-  try {
-    new URL(value);
-    return true;
-  } catch (error) {
-    return 'Incorrect URL';
-  }
-};
